clean and real options for streams and met
