+++ clear-lake-restoration-app/src/Components/pages/data-archive/DataArchive.css
@@ -5,6 +5,7 @@
 
 .archive-container {
     display: flex;
+    flex-direction: column;
     flex-wrap: wrap;
     justify-content: center;
     gap: 25px;
@@ -35,4 +36,31 @@ center > * {
 
 .react-datepicker-wrapper {
     width: fit-content !important;
+}
+
+.clean-real-button {
+    margin: 0;
+    width: fit-content;
+    padding: 0;
+    padding-left: 10px;
+    padding-right: 10px;
+    border-radius: 1.25rem;
+    background-color: #fff;
+}
+
+.center-container {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+}
+
+.button-container {
+    display: flex;
+    flex-direction: row;
+    gap: 5px;
+}
+
+.clicked {
+    background-color: #022851;
+    color: #fff;
 }
